quantum-e-commerce/BE/project/signatures/dilithium_utils.py
# BE/project/signatures/dilithium_utils.py

import hashlib
import json
import base64
from typing import Dict, Tuple, Optional
import numpy as np
from Crypto.Random import get_random_bytes

class DilithiumSignature:
    """
    Simplified Dilithium signature implementation for demonstration
    Note: This is a simplified version. For production, use a proper implementation.
    """
    
    # Dilithium parameters (simplified)
    n = 256
    q = 8380417
    k = 4
    l = 4
    eta = 2
    tau = 39
    beta = 78
    gamma1 = 2**17
    gamma2 = (q - 1) // 88
    omega = 80
    
    @staticmethod
    def generate_keypair() -> Tuple[str, str]:
        """Generate a new keypair"""
        # Simplified key generation
        seed = get_random_bytes(32)
        
        # Generate matrix A (simplified)
        A = np.random.randint(0, DilithiumSignature.q, 
                              (DilithiumSignature.k, DilithiumSignature.l))
        
        # Generate secret vectors (simplified)
        s1 = np.random.randint(-DilithiumSignature.eta, 
                               DilithiumSignature.eta + 1, 
                               DilithiumSignature.l)
        s2 = np.random.randint(-DilithiumSignature.eta, 
                               DilithiumSignature.eta + 1, 
                               DilithiumSignature.k)
        
        # Compute public key (simplified)
        t = (A @ s1 + s2) % DilithiumSignature.q
        
        # Serialize keys
        private_key = {
            'seed': base64.b64encode(seed).decode(),
            's1': s1.tolist(),
            's2': s2.tolist(),
            'A': A.tolist()
        }
        
        public_key = {
            't': t.tolist(),
            'A': A.tolist()
        }
        
        return (
            base64.b64encode(json.dumps(private_key).encode()).decode(),
            base64.b64encode(json.dumps(public_key).encode()).decode()
        )
    
    @staticmethod
    def sign(message: str, private_key: str) -> str:
        """Sign a message with private key"""
        # Decode private key
        try:
            private_key_data = json.loads(
                base64.b64decode(private_key).decode()
            )
        except:
            raise ValueError("Invalid private key format")
        
        # Hash the message
        message_hash = hashlib.sha3_256(message.encode()).digest()
        
        # Generate signature (simplified)
        nonce = get_random_bytes(32)
        
        # Simplified signature generation
        signature_data = {
            'message_hash': base64.b64encode(message_hash).decode(),
            'nonce': base64.b64encode(nonce).decode(),
            'z': np.random.randint(0, 100, DilithiumSignature.l).tolist(),
            'c': base64.b64encode(get_random_bytes(32)).decode()
        }
        
        return base64.b64encode(json.dumps(signature_data).encode()).decode()
    
    @staticmethod
    def verify(message: str, signature: str, public_key: str) -> bool:
        """Verify a signature with public key"""
        try:
            # Decode signature and public key
            signature_data = json.loads(
                base64.b64decode(signature).decode()
            )
            public_key_data = json.loads(
                base64.b64decode(public_key).decode()
            )
            
            # Hash the message
            message_hash = hashlib.sha3_256(message.encode()).digest()
            stored_hash = base64.b64decode(signature_data['message_hash'])
            
            # Verify hash matches
            if message_hash != stored_hash:
                return False
            
            # Simplified verification (always returns True for demo)
            # In real implementation, this would perform actual verification
            return True
            
        except:
            return False
    
    @staticmethod
    def hash_transaction(transaction_data: Dict) -> str:
        """Hash transaction data for signing"""
        # Sort keys for consistent hashing
        sorted_data = json.dumps(transaction_data, sort_keys=True)
        return hashlib.sha3_256(sorted_data.encode()).hexdigest()